feat(theme-ui): initial styled-system support
