+++ packages/docs/src/theme.ts
@@ -36,6 +36,7 @@ const borderWidths = {
 };
 
 const theme = {
+  breakpoints: ["40em", "52em", "64em"],
   colors: {
     ...colors,
     ...colorModes,
